64bit fixes

+++ cocos2d/Platforms/iOS/CCDirectorIOS.h
@@ -61,7 +61,7 @@
 -(BOOL) enableRetinaDisplay:(BOOL)enableRetina;
 
 /** returns the content scale factor */
--(CGFloat) contentScaleFactor;
+-(float) contentScaleFactor;
 @end
 
 #pragma mark -
@@ -99,6 +99,6 @@
 @end
 
 // optimization. Should only be used to read it. Never to write it.
-extern CGFloat	__ccContentScaleFactor;
+extern float	__ccContentScaleFactor;
 
 #endif // __CC_PLATFORM_IOS
